+++ src/app/modules/list/list.service.spec.ts
@@ -1,4 +1,7 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
+import { RouterModule } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { ListService } from './list.service';
 
@@ -6,7 +9,13 @@ describe('ListService', () => {
   let service: ListService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [
+        HttpClientTestingModule,
+        RouterModule.forRoot([]),
+        RouterTestingModule,
+      ]
+    });
     service = TestBed.inject(ListService);
   });