+++ src/app/shared/components/confirm-dialog/confirm-dialog.component.spec.ts
@@ -1,14 +1,39 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MatButton } from '@angular/material/button';
+import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { TranslateFakeLoader, TranslateLoader, TranslateModule } from '@ngx-translate/core';
 
 import { ConfirmDialogComponent } from './confirm-dialog.component';
 
+class DialogRefMock {
+  close() {
+    return true;
+  }
+  getState() {
+    return true;
+  }
+}
 describe('ConfirmDialogComponent', () => {
   let component: ConfirmDialogComponent;
   let fixture: ComponentFixture<ConfirmDialogComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ConfirmDialogComponent ]
+      imports: [
+        TranslateModule.forRoot({
+          loader: {
+            provide: TranslateLoader,
+            useClass: TranslateFakeLoader
+          }
+        })
+      ],
+      declarations: [ 
+        ConfirmDialogComponent
+      ],
+      providers: [
+        { provide: MAT_DIALOG_DATA, useValue: {} },
+        { provide: MatDialogRef, useValue: new DialogRefMock() }
+      ],
     })
     .compileComponents();
   });
@@ -22,4 +47,15 @@ describe('ConfirmDialogComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it('onDismiss', () => {
+    component.onDismiss();
+    expect(component.dialogRef.getState()).toBeTrue();
+  });
+
+  it('onDismiss', () => {
+    component.onConfirm();
+    expect(component.dialogRef.getState()).toBeTrue();
+  });
+
 });